Update teams-calling-policy.md
I had an encounter with a customer who was concerned that enabling the "Busy on Busy" setting to "Unanswered" would prevent calls from being forwarded to the next person in the call queue or sent to voicemail. After conducting tests, we confirmed that this setting does not disrupt the call queue flow and the caller is directed to voicemail. To help others who may encounter a similar situation, I suggest adding a note to the "Unanswered" section/explanation clarifying this information.  Below here is my proposed change to the documentation.  Lines 107 - 110 remain unchanged.  The last sentence my proposed change and would be entered on Line 111.  Another suggestion would be to create a Note box to draw further attention to it.  If there is already existing documentation explaining this, an article reference would also suffice.

- **Not enabled** No busy option is enabled and new or incoming calls can still go to the user while the user is already in a call.
- **Enabled** New or incoming calls will be rejected with a busy signal.
- **Unanswered** The user's unanswered settings will be used, such as routing to voicemail or forwarding to another user.
- **User controlled** Users can determine their busy options choice from call settings in the Teams app. 
Note - Enabling the "Busy on Busy" setting to "Unanswered" does not interfere with call queue operation. If a user sets "Busy on Busy" to "Unanswered," the call will be routed to the next user or sent to voicemail, depending on the call queue's configuration.